use xdg-directory specification for file paths
For backwards compatibility check if the old directories or files exist
before switching to the new location specified by xdg. Use xdg-data-home
for cookies, the bookmark and history dbs, while using xdg-config-home
for init.lisp. Note that we don't need to ensure that
xdg-config-directory exists because it is only used for storing
configuration.