create deferred variables macro
deferred variables are ones that are calculated only during execution
time, this is used as a way to avoid the issue of saving lisp images
with hard coded paths based on my system. this occurs because
variables are top level, and their values get evaluated when the
system is loaded. this solves that issue