NEXT/ENGINE prototypes a pure CL download manager
Implemented a simple proof of concept synchronized request as
NEXT/ENGINE:RESOLVE.  We will use LPARALLEL to orchestrate separate
download tasks in an (eventually) resource constrained manner.

Use a single package for all new symbols for the moment; introduce
dependencies in ASDF subsystems.  Use symbols as symbols, as
preserving string case is not worth the hassle.

Introduce simple declarative tests via PROVE, which may be invoked via:

    (asdf:test-system :next/t)

The use of PROVE as a testing framework requires ASDF to be able to
locate dependencies which can be sticky in some situations, so need to
not "intrude" on the base NEXT system definition by using ASDF
subsystems liberally.

TODO: non-working "alternative code" for determining the PATHNAME NAME
      and TYPE components needs to be re-thought and removed.