Merge define-parenstatic into define-parenscript
The generated parenscript function must support the 3 default arguments together
with user-defined arguments.  The most convenient way to do this is to make all
arguments key arguments.  In particular, this makes it possible for the user to
define parenscript functions with optional arguments.

The WITH-RESULT macro used to expect the callback in first position.  We had to
change it so that now it call the callback found in the :CALLBACK key argument.

All continuation-passing-style code now relies on this :CALLBACK key argument.