Turn commands into mode methods
The main objective of this change is to allow per-mode commands.

When the current mode inherit from another, we want the commands of both modes
to be available.  It allows us to have implement the concept of global commands
vs. local commands.

We leverage CLOS all the way: since modes are classes, it's enough to implement
commands as methods.  This gives us multiple inheritance for free.

As a bonus, it also gives us "hooks" via the auxiliary methods :before, :after
and :around.  This advantageously replaces our former (broken) hook system.

Forthermore, this enforces the distinction between commands and other
functions so that now key sequences can only be bound to commands.

The main drawback, for now, is that in Lisp code commands must be invoked over
their mode, even for global commands.  For instance

  (my-command (mode (active-buffer *interface*)))