ports/gtk-webkit: Hard-code minibuffer height
For now the minibuffer height never changes due to a bug in the port with
GTK/WebKit.
The height is frozen to the first value it is given.  If we echo something
during init, this first value is that of an echo (e.g. 25) which is too small
to display candidate lists.