Rename "s" modifier to "Shift"
We spell it in full to improve clarity and remove possible confusions with
"S" (super).
We could also have spelled "shift" or "SHIFT", but we chose "Shift" for
consistence with other keys like "Lock" or "Right".

The rule of thumb is that special keys are spelled in full in title case.

"C", "M" and "S" are reduce to letters for a few reasons:

- They mean different think on different keyboards (e.g. "M" is often "Alt" on a
Microsoft keyboard while it it "Option" on an Apple keyboard.)

- They are used very often and "Control-x Control-c" is more verbose than "C-x C-c".