Don't set the active buffer when returning from the minibuffer.
This code was present because initially we used to set the minibuffer as the
active buffer, but this was dangerous so the logic was changed.
Now the active buffer is not changed when the minibuffer is simply brought up,
so there is no need to revert the active buffer upon return.