Overhaul use of command class.
The command class has additional slots:
- `bindings' that replaces the `%%command-key-bindings' hash table (easier and
  faster).
- `mode' for convenience.
- `access-time' so we can sort by commands by recency of the call in the minibuffer.

The `object-string' method now displays all bindings of a command instead of
just the latest.

`execute-command' used to be broken in how it found the mode associated to a
command.  This is now fixed.

Many redundant and obsolete functions have been removed (command-symbol,
all-available-commands, etc.)

Know issue:
When calling `define-key` over a keymap, the `:scheme` argument is used to
register the binding in the command object.  This is not very dynamic: if we
manually add a keymap instance to a mode, the command `bindings' slot won't be updated.