Avoid resolving symlinks when listing password-store entries
Resolving symlinks means that we get the full path for the actual file instead
of the path of the symlink. Since `list-passwords` truncates the paths by
droppping the length of the password-store path from password files, this means
that we'll drop arbitrary parts of the path name for symlinked entries.

Consider this tree

~~~
~/.password-store/www/somesite.com.gpg
~/.password-store/www/anothersite.dk.gpg
~/.password-store/work -> ~/work/passwords                            # symlink
~/work/pass/www/mywork.example.org.gpg
~~~

If we truncate all entries with the length of the primary password store
prefix (`~/.password-store/`), we'll get these password entries:

~~~
www/somesite.com
www/anothersite.dk
work.example.org
~~~

...which is not quite what we want. If we don't resolve symlinks we'll start
with this:

~~~
~/.password-store/www/somesite.com.gpg
~/.password-store/www/anothersite.dk.gpg
~/.password-store/work/www/mywork.example.org.gpg
~~~

and end up with this:

~~~
www/somesite.com.gpg
www/anothersite.dk.gpg
work/www/mywork.example.org.gpg
~~~