delete pyqt-webengine
Woe is me, O Hector; woe, indeed, that to share a common lot we were
born, you at Troy in the house of Priam, and I at Thebes under the
wooded mountain of Plakos in the house of Eetion who brought me up
when I was a child – ill-starred sire of an ill-starred daughter –
would that he had never begotten me. You are now going into the house
of Hades under the secret places of the earth, and you leave me a
sorrowing widow in your house. The child, of whom you and I are the
unhappy parents, is as yet a mere infant. Now that you are gone, O
Hector, you can do nothing for him nor he for you.