Add link/button hint and matches follow-focus
This makes link/button-hints and search results be able to be
followed with M-n and M-p which are mapped to the two new commands
select-next-follow and select-previous-follow.

Because M-p was previously mapped to minibuffer-history,
minibuffer-history has been remapped to M-h.

When you select a result in the minibuffer
with either of the new commands, the selected result will be
highlighted, the buffer will scroll it into view, and if it is on
another buffer than the current one, which can be the case with
multi-buffer searches, a switch will happen to the correct buffer.

This also changes the the C-s shortcut that was available for search
results which will focus the currently-selected result. Now it uses
the same feature as above, so it will also get the highlight.