minibuffer changed callback
This adds a callback to the minibuffer that gets called every time
the minibuffer has been updated. This is useful for when you e.g.
want to react to changes in the minibuffer that should be reflected
in a buffer, like for example when highlighting the currently selected
search result.