minibuffer.lisp: do not clear input-buffer immediately
setup-default to clear the input-buffer was necessary when we had a
single minibuffer object. now that we make instances of the
minibuffer, it is not necessary to clear it, only to clear the
web-view that it will use. in order to invoke the function to generate
the completions etc, we have to invoke setf, probably we could pull
out the relevant parts from the setf definition and make things cleaner