next.asd: Simplify and normalize test calls.
Previous calling `asdf:test-system' over FOO instead of FOO/TESTS would do
nothing.

prove-asdf is only useful for :test-file if we want to declare tests in the same
system, which Guix does not understand.  It's not very idiomatic anyways, let's
not do it.  This also makes it simpler to load the ASD since it does not have
dependencies anymore.