Decode all URL displays.
Hex-encoded URLs are not human-readable, e.g.
https://ta.wikipedia.org/wiki/%E0%AE%B5%E0%AE%BF%E0%AE%95%E0%AF%8D%E0%AE%95%E0%AE%BF%E0%AE%AA%E0%AF%8D%E0%AE%AA%E0%AF%80%E0%AE%9F%E0%AE%BF%E0%AE%AF%E0%AE%BE.

We do not decode URLs internally, e.g. we keep http://foo/bar%20baz encoded as
is, but we display "http://foo/bar baz".

The only trick is with tab-completion: we can't insert the object-display
because it contains too much information (e.g. the title in "URL title"), so
instead we keep inserting the object-string and if it is a URL then we decode it.