Restore sessions lazily.
Previously we would tell GTK to load all restored buffers on startup.  This
would hang the browser for a while before it displayed the first window because
the queries are done within a single GTK main loop cycle.

It's hard to load more than very few URIs in a single GTK loop cycle because of
how long it takes.

The easiest (and possibly best) option is to restore the buffers on demand.
This is what this patch does.

Another option would be to restore about one buffer per second, or when the
browser is idle.  The benefit is that this would make all the buffers ready to
browse after a certain time and the user would not have to wait on every buffer
switch.