Fall back to default URL instead of broken replacement buffer.
The replacement buffer typically has no mode, which prevents the user from
using any key binding (save the bindings from the override-map).

A better approach is to load the default URL, after all it's what it's made for.
To load nothing (which simulates what we used to have), it's enough to set the
default URL to "about:blank".