Fix window-set-active-buffer.
This was easily broken before with just 2 windows and swapping buffers.
The reason is that we were calling some high level functions instead of the low
level ones and vice versa.

Now that the temp-buffer is a low-level buffer instance that is not registered
in the browser buffers, there is not need for the (get-inactive-buffers)
cleanup.