Allow quri:uri in set-url* and rename it to buffer-load.
set-url* was a wrapper around the low-level ffi-buffer-load, so to be consistent
with out other wrappers, we rename it like its low-level function minus "ffi-".

Now that our URLs are quri:uri objects, we can easily make the distinction
between raw URLs and user input that needs to be passed to parse-url: we simply
check if the argument is a string or a quri:uri.
This rids us of all the object-string conversions.

While we are at it, we unexport set-url-hook which was mistakenly exported.  We
also rename it to buffer-load-hook for consistency.