libraries/history-tree: Update node data on match.
Previously, if we tried to add a an already present URL to the buffer history,
the new buffer-description would be ignored.  This prevented the title from
being set up properly since the notify::title signal is sent after the
notify::url signal.

Now when the entry already exists, we update its data.