Define default search engine via default-search-engine slot
Previously, the magic value "default" in a search-engine signified it as
the default search engine. This commit factors that hard-coded value
into the browser slot default-search-engine instead.

The motivating use case for this change is to provide modes an easy
method of changing the default search engine. In particular, users may
wish to default to an onion URL when proxy-mode is enabled with Tor.