help: Fix class slot documentation in describe-class.
Previously we used to define a "public slot" as a slot that has at least a
reader or a writer.  This is inconvenient because closer-mop has a
slot-definition-readers method only for direct slots.

Anyways, this rule may be too confusing and it's simpler to just consider a
"public slot" one that is exported.