class-star: Remove superclasses cycle detection.
This allowed us to override a class by itself, e.g.

(define-class foo (foo)
  ((new-slot ...)))

This is brittle however: typing is not well specified then on the various
implementations.
The children of FOO might not be of the new FOO type.
This may also break method dispatch.

Overall, it's a bad idea.