Restore user-data just-in-time.
Previously, user-data was only restored when with-data-access was called.
This caused the first buffer to have no history and bookmarks would not be
loaded until a bookmark was saved, which erased everything.

Instead, we restore the data in get-user-data to ensure it's always loaded.

This also fix a second bug which caused the data to be restored every time
with-data-access was called, while there is only need to restore the data once.