help.lisp: add type checking to configure slot
Type checking is optional, if you pass in a type to configure-slot,
checking will be done. Currently there is a problem with echo. If you
try to do the following:

(if (typep (read-from-string input) type)
                             (return input)
                             (echo "Incorrect type submitted"))

the program will show an echo and dismiss the minibuffer. What we
would like to do is show the echo and NOT dismiss the minibuffer. So,
for now, just keep prompting the user until the type is correct.