text-buffer.lisp: Added new approach to word-wise cursor movements
Introduced a new global variable that customizes the behaviour of
cursor-forwards-word and cursor-backwards-word.

By default, the cursor navigates to both left and right boundaries of
words.

Without loss of generality, let's look at the case of
cursor-forwards-word.

Let the symbol "|" denote the position of the cursor after sucessive
calls of cursor-forwards-word.  Assume that the cursor started off at
the beginning of the line:

a |piece| |of| |text|

When *conservative-mode-word* is non-nil:

a| piece| of| text|