buffer: Use current-history-node to control history branching.
The problem was: how do we make sure that new buffer starts its
history from the node of parent buffer that was current at the time of
its creation? Modifying the history tree is not an option there. In
the end, we're just storing the parent history node in the child
buffer.