Add ensured default value to with-data-access.
This adds a new keyword parameter to with-data-access. This parameter
specifies what value to assign to the given variable if data is not
found.

One important addition is that all the arguments of with-data-access
are now parenthesised -- to ease argument parsing. This, however,
looks similar to let-bindings and other with-* macros, so it shouldn't
be hard to parse.