libraries/history-tree: Overhaul the classes to welcome the global history tree.
Previously, the history-tree structure was meant for single actors and would not
be shared between multiple actors.

This is limiting if we want to implement a global history tree, that is, a
history tree that's shared between multiple actors.

This has a few benefits, among others:

- Any actor can navigate to the history trees of any other actor, which allows
for going beyond the local tree.

- It preserves the relationships between the actors, that is, from which node of
which actor a new actor is spawned.