Revert "libraries/history-tree: Draft reparenting helpers."
This reverts commit cec1d654a9598f8452987ae3a8cb1e50bbc16e46.

This is too complicated, not many users are ever going to grasp this.

It's simpler to keep the histories immutable.  If the user wants to delete en
entry, instead they can should call a function, say "reset-history" over all the
owners, which leaves the owners with only the current node, thus in effect
freeing all entries.

The result is that most of the entries in the entry stash (except those pointed
by the new current nodes) are available for deletion.