libraries/history-tree: Store entry last access in a new entry-accessors class.
We need this because if an entry has no node, then we would have no fast to query
the last access just from data.  If we had stored the last access in the `entry'
object itself, we would have to loop over all entries until it matched the DATA
argument.
This is way too slow.