Don't load history for first `help'  buffer.
Previously, we didn't register the history of any internal buffer anyways, so in
practice this makes no change history-wise.  But at least it allows us to render
the help page, since history restoration is done only afterwards in the default
startup function.

In the future, we would like to enable history for internal buffers as well,
which means that calling the `help' command should prompt a buffer with history
support.