Allow for interactive password-interface fixing.
This introduces a questionable fill-interface generic (it's
questionable, because its methods need to be in source/password.lisp
instead of libraries/password-manager/password-*.lisp) that fixes all
the fields of the interface in case the password in wrong.

Is implemented for keepassxc-interface only, but can be useful to the
future ones (lastpass etc).