Remove startup-function slot: require explicit overwriting of startup method
We need more control over the startup function. The startup-function
slot creates a level of indirection and difficulty in debugging. we
rely on too many things to occur within a normal startup. Instead of
creating a slot that the user can change the value of, we require the
creation of a complete method to override the old one. This is a more
drastic change and requires the user to actually inspect/understand
the code for the startup.