Make modes rememberable.
This adds rememberable-p slot to root-mode and (all) the inheriting
modes. It's set to nil for those modes that we don't want to be
remembered by auto-mode, like internal-buffer styling modes or
one-shot modes, like visual-mode or expedition-mode.

This, however, shifts some of the responsibilities of auto-mode back
to modes, instead of universal mode-invocations. Should it be that
way? I'm not sure.