Fixes xdg-open from using the imporper application
Currently nyxt uses namestring which turns a downloaded file to a name
like file:/home/user/Downloads/a.pdf. This causes xdg-open to use
x-www-browser instead of the extension. This change causes the
application to instead use file:///home/user/Downloads/a.pdf, in which
xdg will use the proper application