Auto-switch to vi-mode when focusing on a focusable element.
Add element-focused method to allow modes to customize the behaviour when an
element is focused.

Replace focus-element parenscript with a macro that handles the call to the
element-focused.

Customize the behaviour for vi-normal-mode to switch to insert mode.
It works for focus-first-input-field and all element focus functions.

Remove vi-focus-first-input-field since it's no longer needed.

Move focus-first-input-field to web-mode (where it should be).

Load `message' before buffer, etc. since it does not depend on much and echo
functions are needed everywhere.