Fix invalid .desktop entry
Remove an Actions entry that doesn't have a corresponding action group
and remove version entry since the old value was invalid (version is
supposed to refer to the version of the desktop entry file, not the
version of the application) and the key is optional.

https://specifications.freedesktop.org/desktop-entry-spec/latest/ar01s06.html

---
Version of the Desktop Entry Specification that the desktop entry
conforms with. Entries that confirm with this version of the
specification should use 1.5. Note that the version field is not
required to be present.
---

https://specifications.freedesktop.org/desktop-entry-spec/latest/ar01s11.html

---
Each action is identified by a string, following the same format as key
names (see the section called “Entries”). Each identifier is associated
with an action group that must be present in the .desktop file. The
action group is a group named Desktop Action %s, where %s is the action
identifier.
---