libraries/history-tree: Don't replace entry data on match.
This radically changes the behaviour of the history to something much more
useful: instead of replace the matched data, we keep the old one.

This allows the caller to persist data and act upon it from the call site.
For instance, in case of URLs, the caller can decide to update the "title" or
not.