history.lisp: on bookmark-frequently-visited-urls command, insert a `if-confirm` macro. There are two advantages.
(i) The user will be asked if he/she really
wants to bookmark a frenquelty visited which prevents undesirable
bookmarks; and,

(ii) The question brings awareness about which page
is being bookmarked. Hence, the choice of the tags will be easier.
Before this addition, the GUI was not telling the user which page was
being bookmarked.