Don't use dirname for init-data-path, auto-init-data-path and socket-data-path.
Otherwise setting the dirname in the expand-data-path method would change the
location of the files, i.e. the location would change after loading an init file
which defines a nyxt:expand-data-path that sets the dirname of the data paths.

We only want to be able to set the location of the aforementioned data paths
from command line arguments.