data-storage: Add data-path-source.
We register all instantiated data-paths.  To avoid collecting them
indefinitely (e.g. after their owner is garbage collected), we register them in
a weak hash table.

For this we need the trivial-garbage library which is already an indirect
dependency.

We also add an `editable' slot to `data-path' to avoid listing paths like the
socket.