prompt-buffer: Fix C-v, etc. not updating the suggestions.
Previously, when the prompt buffer HTML input was changed without pressing a
key, the suggestions were not updated.

From now on, we should only edit the HTML input via set-prompt-buffer-input,
which will ensure that the suggestions are updated.

We seize this opportunity to remove the rather useless and confusing
set-prompt-input.