Remove duplicate about `quit` command
`quit` command appears two times in the tutorial on the Miscellaneous and the Quick Start sections. Not sure if this was intentional. I guess it is just a small accident. Thus, I am submitting a patch to fix it. I am keeping it on the Quick Start since it is a crucial binding.