gtk.lisp: Web context sharing and separation
Prior to this commit, nyxt would create a new web-context for each
buffer, but back each non-ephemeral web-context to one of two
locations on disk, ("data-manager" for external buffers, and
"WebKitCache" for internal buffers (as the webkit-data-manager wasn't
being set, so would default)). Since they were backed to the same
location on disk, some session-sharing was possible, but in an
unreliable manner. Furthermore, while the data-directory was set to be
inside of xdg-cache-home the cache-directory wasn't being set, so some
caches were being stored inside of the data-manager-data-path and
others on the system-wide webkit default (hsts-cache-directory and
offline-application-cache-directory).

This patch causes nyxt, by default, to re-use in-memory web-contexts
based on if the buffer is internal or external and whether or not it
is meant to be ephemeral, determined by the make-web-view
function. This function can now also return a web-view in a custom
web-context.

Data-path names are generated based on the name of the context and
cookies are now stored separately per context. The data-manager's
data-directory is stored inside of xdg-data-home and its
cache-directory is stored inside of xdg-cache-home.

As buffers are now assigned to a web-context, it doesn't make sense to
store defaults controlling web-contexts in buffer classes. The
default-cookie policy was moved to a setting in the browser and a
per-context cookie data-path is generated

This sets the groundwork to allow user-facing browsing "sessions",
akin to Firefox's "Multi Account Container"s, where users can assign
groups of buffers to share a browsing context.