Pass extensions-loadable resources as data: URLs.
The reason for that is that custom schemes don't allow loading
resources in WebKitGTK (WebKit itself allows for that, but there are
no GTK bindings for that way yet). Thus, we need to bypass it with the
help of data: URLs. Not exactly safe and reliable, but should work in
most cases.

This also adds a runtime.name property to easily identify extension by
name.