gtk.lisp: create outside-renderer-thread
To run javascript and return it successfully it has to be run on a
thread outside of the main thread, that is because there is a callback
that the user must wait for. Incidentally, this callback ALSO runs on
the main thread. Thus, we must use outside-renderer-thread for some
operations that wish to retrieve values from the renderer itself.