web-extensions.lisp,renderers/gtk.lisp: Add Lisp-side permissions.
This is preferable to be done on the C-side, but it's too complex to
do there. Sometime later, maybe.

This is not complete: tabs-query and tabs-create should also check
extensions for permissions, because they return tab descriptions, and
descriptions should not include url and title when done for extensions
that have none of suitable permissions. This, however, requires us to
somehow identify extensions in _every_ WebKitUserMessage, which asks
for quite a big architectural change in the C library.

This also includes stricter typing of all the Lisp-side API callbacks,
some minor bugfixes and reorderings.