buffer: Fix race condition when deleting buffers.
The race condition would happen when deleting many buffers (30+, depending on
your hardware and the 'weigth' of the web page).

We alter the history in only one spot: ffi-buffer-delete is run in a different
thread and it used to also alter the history, which would lead to a race
condition.
Now that all the history changes are done at the same spot on the same thread,
we should be safe.