Replace buffer-make by `initialize-instance :after'.
This simplifies the API:
- It removes the confusion between `buffer-make' and `make-buffer'.
- It allows for the caller to pass any of the buffer class initargs to
`make-buffer', or just call `make-instance 'user-buffer' since this now results
in a valid buffer object.

The main collateral change is that now dead buffers are resurrected via the
dedicated function `ressurect-buffer' instead of calling `buffer-make'.
Semantically this makes more sense plus it simplifies the logic both for buffer
creation and dead-buffer resurrection.