repl: Generate evaluation history when loading the lisp-repl URL.
This gets rid of bugs with buffer reload visually clearing the
evaluation history. It also integrates REPL better with the internal
pages infrastructure.

The downside is that to update it we need to either
- Reload the buffer.
- Fill the page with JS (as we were doing before).

This patch sticks to the former, while latter may have in fact worked
better. It's not that anoying, given that the load is almost
instantaneous.