GTK: Fix download page not being refreshed when downloading from it.
The current URL is that of the download, while the queried URL is
nyxt:list-downloads, which is what the current buffer _displays_.

Previously that would trigger the code to fetch the page content from history.
The trick is to simply not use WebKit cache for internal pages, it's not useful anyways.