dom: Add get-unique-selector utility to identify page elements.
This allows us to attach annotations easier and find elements with more
precision. Maybe we can even ditch nyxt-identifier attribute and rely on unique
selectors? For this we need to improve the selectors and ensure they are really
unique (which is not always possible, I suppose...)

This also adds `parents' method to get a list of recursive node parents.