Ensure prompter is destroyed after it's selection is returned.
This is important for use in `query-hints` since without ensuring the prompt is
destroyed the hints will remain on the screen. This is only an issue when the
hints are used to select a input window since this does not refresh the buffer
and therefor will never remove the hints.