Instead of destroying the prompter simply remove the hint elements
`resume-prompt` will now function properly while removing the hints.
The trade off is the fact that these hints will still not be restored.

Would it be best to prevent hint commands from being resumed?

Should we simply create new hints?

Should we attempt to reproduce the exact same hints in the same buffer at the
same url? This option is likely a little overkill but should be possible.