Revert "foreign-interface.lisp: create package nyxt/ffi"
This reverts commit d52c9de5a320f8524d4604c0224b43212a418f52.

The benefits don't seem to outweigh the cost.
To avoid all these forward references, we would have to define the default
methods after all the core files are loaded.

Which means, split the FFI interface into 2 files.  Not ideal...

Besides the current state is broken, it does not render anything.