Garbage-collect Lisp URL callbacks.
Storing them in a global table means they would never be garbage-collected.
A weak hash-table won't do: there are no references to the callbacks, so they
are garbage-collected too soon.

Instead we store a hash-table per buffer and we clear it on load since the Lisp
URLs are then erased from the content of the buffer.