Define command manually for internal pages using non-scalar parameters.
This is often the case of internal-pages that rely on the buffer ID.
Instead, we can define the command ourselves of the BUFFER object and pass the
ID to the internal page.

This makes command more consistent and saves us from exposing an implementation
detail at level of command parameters.