browser: Always restore session by default.
Also turn session-restore-prompt into the restore-session-on-startup-p boolean.

This greatly simplifies the startup logic which tended to always break.
Besdies, prompting the user for which session to restore is a rather advanced
use case.  Makes little sense to default to this for newcomers.

It's simpler if the user restores the session manually.