buffer(finalize-buffer): Turn extra-modes into a list too.
This caused massive problems when using `prompt :extra-modes`, given that
`customize-instance :after` was properly processing :extra-modes into a list,
while finalize-buffer wasn't, causing the mode symbol to become cdr of the list
and misshaping it to the point enable-modes was unable to cope with it.

Do we actually need `customize-instance :after` on prompt-buffer?
finalize-buffer seems to do the same job of setting :extra-modes...