Add toggle-maximize command
This commit adds a command to Nyxt to maximize and unmaximize a window. This
requires a change in cl-webengine to expose widgetShowMaximized to Common Lisp
as widget-show-maximized. I have duplicated ffi-window-unfullscreen on the QT
side as ffi-window-unmaximize to keep things symmetrical with GTK (GTK has a
specific function to unmaximize a window while QT just has "showNormal").