configuration(define-configuration): Warn about :around methods.
Those are usually generated when we're clueless about the slot/method being
configured, and thus are red flags of their own.

Maybe even remove this clause and allow Nyxt to err out, so that methods are not
being wrapped at all? But still, define-configuration may be useful even for
those classes/methods that we're not explicitly allowing to configure...