Move download hooks to nyxt/download-mode:download.
This way, the type for hooks is defined at the right time for them to work.

I've also cleaned up and uniformized the places hooks are being called in. In
particular, (nyxt/download-mode:)before-download-hook is not invoked right after
setting the cancel function, allowing one to cancel the download just after it
started.

Closes #2512.