urls: Set the URL of the internal-page-loading buffers.
Closes #2554.

The problem is that our logic for internal pages, with the removal
of (current-buffer) hack in df55622fd08273152b2eb5de963d527be1d7831f and
b8a34c69cf135e0cf5abbe35400e5c13768e34fa, have lost an ability to find the
buffer that the nyxt: URL is loaded in. The buggy sequence was:
- (Nyxt) buffer-load the nyxt: URL.
- (Renderer/Nyxt) call the internal-page function.
- (Nyxt) internal-page function calls find-internal-page-buffer to get the
  buffer.
- (Renderer) don't emit "notify:uri" signal until "nyxt:" URL is done.
- (Nyxt) don't (setf url) of the buffer, because renderer doesn't tell to.
- (Nyxt) find-internal-page-buffer doesn't find the right buffer,
  - because the right buffer doesn't have the URL set yet,
    - because renderer still waits for the "nyxt:" URL to load.

So, as a solution to this vicious cycle, we (setf url) of the right buffer just
before applying the internal page function. This should be safe even in case
internal page function will fail, because if it will, return nil and open the
error page then.