mode(resolve-symbol): Speed up, don't memoize.
The initial reason for memoization was that resolve-symbol was hella slow when
searching across (list-all-packages). The reason was an O(N^3) algorithm for
subpackage search. I've sped it up to O(N^2) and pruned the list of possible
subpackages.

Relates to the caching problem from #2540, but doesn't yet solve it—describe-*
pages still have NIL page-mode, because help-mode doesn't yet exist when those
are defined.