mode/hint: Refactor deduplication logic.
The old method is too expensive and it sacrifices the interactive nature of the
prompt buffer.

For a page with 255 hints, it takes 1.5 seconds for the deduplication to finish
on a reasonable machine.  Notice that pages can have more than 2000 hints.

The new method takes ~0.7 seconds on a page with 2000 pages.