renderer-script(setf form): Preserve the original type and status.
The problem with the old implementation was that one couldn't return, e.g.,
"image/jpeg" type and see it in the browser.

This commit also alters the HTML pages only in the cases when there's no <head>
tag. If this tag is provided, it's implied that the page author knows what
they're doing. In case of mode/repl.lisp, the internal page is altered to not
trigger this behavior.