describe(describe-any): Turn into an internal page.
This way, we fix broken JS links in resolve-backtick-quote-links and allow
invoking describe-any as both a prompting command to find _anything_, and as an
internal page to show the disambiguations for any ambiguous symbol.

I don't like this

(run-thread ""
 ...
 (buffer-delete ...))
""

hack there. But, until we properly process redirections in cl-webkit (easy, but
I just haven't gotten myself to do it), this is the best way.