spinneret-tags(:nsection): Add a link and allow adjusting the level.
The trick with the hash-anchor is stolen from GTK docs, it's too nice to
ignore. An alternative would be even more conventional chain link in front of
the heading (maybe even as a CSS style), but I find the hash-anchor more
aesthetically pleasing.

The level adjustment is necessary for cases when the Spinneret code is not
generated in one piece, but in many different pieces. Having :nsection to manage
heading level on its own in such cases results in abundance of <h1>, which is
not exactly cool, is it?