nyxt.asd: introduce documentation coverage subsystem
The documentation coverage subsystem nyxt/documentation-coverage creates a file
named documentation-coverage.txt at the root directory of the Nyxt
repository. The file contains statistics on undocumented symbols of internal
packages of the nyxt system. One may view a particular package with

    grep MYPKG: documentation-coverage.txt | fmt

or any other command that searches for the string `MYPKG:` in the text file.