describe(describe-command): Downgrade to calling describe-function.
This way, we don't have the divergence of features between describe-command and
describe-function, while having an intuitively-named command for interactive
use.

The downside is: we can no longer use describe-command as an internal page URL.
But, given that this use was mostly Nyxt-internal, this is solved by knowing
that this is no longer an internal page and avoiding to use it like one.

Closes #2600.