auto-rules: Fix modes not being remembered when on non-rule page.
The problem was: we used to have (dis|en)able-mode-hook handler that'd remember
modes into last-active-modes one by one, and it worked more or less fine. Then I
removed it during auto-mode -> auto-rules refactoring and forgot to provide the
replacement for the case when:
- prompt-on-mode-toggle-p is nil, and
- the mode is simply toggled on or off,
- after which a reload happens.

The solution was to add the code handling exactly this case and remembering
modes into last-active-modes when necessary. Right now it mirrors the old logic,
while looking a bit cleaner :D