gi-gtk/tests: Remove with-headless logic.
The same functionality can be achieved by starting Nyxt with:

(start :headless t)

This was probably introduced in a time when the handling of the global
variable *headless-p* was flawed.  Regardless, it makes no sense to fix it in
the context of testing, but in the context of its origin (i.e. Nyxt's start
logic).