input(dispatch-input-event): Resolve bound symbols to commands.
This is convenient for cases when one binds the key to e.g., 'input-edit-mode
and does not know that the proper name is
'nyxt/input-edit-mode:input-edit-mode. Using a mere 'input-edit-mode used to not
work, because 'nyxt-user:input-edit-mode is obviously unbound.

This commit uses the magic of sym:resolve-symbol to resolve the unambiguous
symbols to the commands when dispatching the user-pressed keys.