renderer-script(set-internal-page-method): Fix internal page args.
The problem was: arguments with default values are not becoming part of &rest
argument list, because they were not provided during function call. Thus, even
if we set them to a reasonable value, they are lost. Which is not cool.

The solution is to fall back to the old provide-all-keywords way. It's not nice,
but it's reliable.

An alternative way would be to actually look at default values and only provide
arguments that either were provided or have a default value, but that's too much
guesses and implementation difficulties already. Sometime later.