urls: Fix the predefined macroexpansions of lisp-eval.
Closes #2760.

So the problem (that made REPL cells redirect things to the first one and mode
clicks in the status area to behave as the clicks for the first mode; in
general, all the loop- or dolist-generated Spinneret code was corrupted) worked
this way:

- Spinneret parses HTML and gets to some :onclick or :onfocus with
  nyxt/ps:lisp-eval.
- Lisp/Parenscript compiler starts expanding nyxt/ps:lisp-eval.
- The id is generated only once, on macroexpansion time, not as a variable local
  to the particular element, e.g.
(dolist (e three-elements)
 (:div :onclick (ps:ps (nyxt/ps:lisp-eval (:title "foo") (foo e)))))
expands to
(:div :onclick (ps:ps (nyxt/ps:lisp-call "1")))
(:div :onclick (ps:ps (nyxt/ps:lisp-call "1")))
(:div :onclick (ps:ps (nyxt/ps:lisp-call "1")))
because id is set to "1" at the time of nyxt/ps:lisp-eval expansion.

Generating the ID on the Lisp-side literally in every piece of macroexpanded
code solves the problem.

That was a hell of a macroexpansion ride, though...