urls(lisp-url-flet): Fix.
The same problem as with lisp-eval in 27612fee394f80dee6480c045ec7da5cd1f82196:
the generated ID was not local to the lisp-url-flet expansion, it was local to
the macro, thus expanding with opaque and same-ish IDs everywhere.

Additionally, the expansion of the local macro was using cons improperly. That's
fixed too.