urls(render-url): Get back the decoded display for the safe URLs.
This way, they take up less space and look much more readable in the
interface. This feature already was on Nyxt, but was randomly removed in the IDN
fix.

Also includes a small refactoring of render-url (more rethinking required, as
it's quite a piece of spaghetti (not without my contribution) right now).