prompt-buffer.lisp: compromise on a tr:hover
we cannot have both a onmouseover event with a onclick event below it, WebKit
does not seem to allow it.

if we have onmouseover, the onclick event never gets fired. This goes for even
things as simple as anchor tags.

thus we have to change the style via hover, and then only enable an "onclick" listener