prompt-buffer.lisp: Add support for ctrl + click to select items.
You cannot pass the event in via the onclick attribute. Even if you do wrap
things like this:

(ps:ps chain ((lambda (event) ..))) the event will never resolve. However, you
can get the event from the global context by using `window.event` the exception
being that this is not supported in the Gecko engine.

Using this event, we can tell if the user has clicked with the control button depressed!