status.lisp: Make tabs static in the UI.
By default, it's preferable for tabs not to be shuffled based on last access
time.  This mimics the behavior of conventional browsers.  Notice that this only
applies to the tab area.

New tabs are always added on the rightmost extreme of the tab area, also
following the behavior of conventional browsers.