libraries/theme: Add alternate background color.
This color is a nuanced version of background-color.

Surface colors, such as primary and secondary, are required to feature a
contrast ratio of 7:1 with respect to their on-color counterparts.

Background colors feature the maximum contrast with respect to their
on-color counterparts (21:1).

Alternate background colors are in-between the 2 above.  They're required to
feature a contrast ratio of 12:1 with respect to their on-color counterparts.