Use filter-exact-matches in much more places.
This way, sources disappear when having no matches, or narrow down to the useful
suggestions really fast. This is particularly important when dealing with lots
of noisy data, like URLs, colors, lots of text entries.

The default preprocessor, delete-inexact-matches, is insufficient, as it matches
all the suggestion against any substring of input, which almost always passes
most elements through.