mode/repl: Prevent the undeletable cell bug.
When the :form argument to the repl command was non-nil, repl was adding the
form after any action, thus cluttering the REPL and staying even after the
"Delete all" button is pressed. Reloading the REPL via the command instead of
reload-buffer solves this problem—all the arguments are purged after the first
use.